+++ spec/classes/init_spec.rb
@@ -12,7 +12,16 @@ RSpec.describe 'fluentd' do
     include_examples 'works'
   end
 
-  context 'with defaults for all parameters', :redhat do
+  context 'with redhat', :redhat do
     include_examples 'works'
+  end
+
+  context 'with plugins', :redhat do
+    let(:params) { { plugins: { plugin_name => plugin_params } } }
+
+    let(:plugin_name) { 'fluent-plugin-http' }
+    let(:plugin_params) { { 'plugin_ensure' => '0.1.0' } }
+
+    it { is_expected.to contain_fluentd__plugin(plugin_name).with(plugin_params) }
   end
 end