+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+ - Add param `plugins`
+
 ## 2016-10-13 - Release v. 0.7.0
 
  - Add param `service_provider` (@larsks)